+++ Route/RouteDefinition.php
@@ -1,0 +1,39 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Bartacus Platform.sh bundle.
+ *
+ * Copyright (c) 2017 Patrik Karisch
+ *
+ * This bundle is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This bundle is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this bundle. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Bartacus\Bundle\PlatformshBundle\Route;
+
+use Psr\Http\Message\UriInterface;
+
+interface RouteDefinition
+{
+    /**
+     * Get the original url, the route key from routes.yaml.
+     */
+    public function originalUrl(): UriInterface;
+
+    /**
+     * Get the resolved url, either a productive domain or a mangled environment url.
+     */
+    public function resolvedUrl(): UriInterface;
+}